:sparkles: remove patcher for setTimeout
